Hanbin: update the levelset method initial point and fault simulation
